requirements.txt
playwright==1.40.0
icalendar==5.0.11
firebase-admin==6.3.0
python-dotenv==1.0.0
